blackmage fixes
try for all levels
